[CS-3533]: Add Change wallet on my wallet address
